Update intro-outro.zss
Fixed a bug that caused the camera to always zoom on p2 during time over. Added additional instances of ignorehitpause to make up for a shortcoming in 0.98.2 (already fixed in the main repository, should be available in 0.99). Removed some additional spacing that didn't really make the code easier to read.